src/python/requirements.txt
hdfs
fastavro
pyarrow
pandas
kafka-python-ng
pyspark==3.5.2
pandas
matplotlib
datetime